refactor: move createUsage to util
